fix: remove AVX instruction from mf-owhm.vfproj
The Visual Studio Visual Fortran Project for mf-owhm
included the option: GenAlternateCodePaths="codeForAVX"

This enabled the AVX acceleration extension if the CPU supports it.
However, part of the acceleration is to truncate double precision variables.
This resulted in model results being slightly different on repeated runs
with the same input parameters.

This option only offered a minor performance increase and was removed
to improve repeatability of simulation runs.