feat: hob obsnam max length is variable
Added keyword "OBSNAM_LENGTH ilen",
where ilen is the maximum length
of any HOB observation name (obsnam).

The keyword is specified at the start
of the HOB input file and must be
the only keyword on the line.

If it is used with TIME_STEP_PRINT or
TIME_STEP_PRINT_ALL, then only one keyword
is specified per line and the order does not matter