fix: sub storage property when h is less than botm
Added code to have the SUB storage coefficient
only release water when HNEW and HOLD is above
the cell bottom.

If HNEW is less than BOTM, then inelastic
compaction is complete and the critical head
is set to -inf, which is -3.4e38

This mimics all water is released from the aquifer
flow package when a layer transitions from
confine to unconfine.

The following code illustrates this:

C3------DETERMINE STORAGE CAPACITIES FOR CELL AT START AND END OF STEP
       RHO1=SCE(LOC)*TLED
       RHO2=RHO1
       HCTMP = HC(LOC)
       HHNEW = HNEW(IC,IR,K)
       HHOLD = HOLD(IC,IR,K)
       BOT   = BOTM(IC,IR,LB)
       !
       IF(HHNEW.LE.BOT .AND. HHOLD.LE.BOT) CYCLE  ! DO NOT MODIFY RHS AND HCOF IF HNEW ARE LESS THAN THE BOTTOM OF MODEL CELL
       IF(HHOLD < BOT) HHOLD=BOT                  ! ADDED CHECK FOR WHEN HOLD<BOT THEN SET TO BOT
       !
       IF(HHNEW < HCTMP) THEN
           IF    (HHNEW >= BOT) THEN
                                RHO2=SCV(LOC)*TLED
           ELSEIF(BOT < HCTMP) THEN
                                RHO2=SCV(LOC)*TLED
           ELSE
                                HCTMP = ZERO
           END IF
       ELSE
           HCTMP = ZERO  ! Note that (RHO2-RHO1) = 0, so HCTMP*(RHO2-RHO1) -> 0*(0)
       END IF