feat: bas option save_head
The BAS package options now incudes one of
the following options:
SAVE_HEAD  LAST_TIMESTEP  generic_output
or
SAVE_HEAD EVERY_TIMESTEP  generic_output

This option writes the entire head array
for either the last time step of every stress period
or for every time step in the simulation.

The output supports the post-keyword
BINARY
for Fortran stream binary format
and the post-keyword
                    SIGFIG  int
which specifies the number of
significant digits (int) to use when writing
the head arrays.