refactor: sub removed fortran list-directed read
SUB package previously used a list-directed read
to load most of the input. The problem with this
is that is missing input is handled differently for
different compilers (typically setting the number to 0).

Also if the input read fails it just results in an
execution failure and stack dump to the screen. The
new read util GET_INTEGER and GET_NUMBER instead
raise a clean error messages and exit the program.

Further, the new read util allow comments on any input line.