feat: cbc writing constant head array is optional
The BAS package now supports the option:
CONSTANT_HEAD_BUDGET_OPTIONAL
that indicates that if the model does not include
constant heads (IBOUND < 0, CHD, or FHB),
then the '   CONSTANT HEAD' array is
not written to the CBC file.

Previously, if constant heads are not part of the
simulation, then each time the flow package writes
to the CBC a there is an NROW by NCOL by NLAY array
of all zeros written.

This option reduces the CBC size by about 5%.
However, it may break some existing tools
that rely on the permanent presence of the
'   CONSTANT HEAD' record name.

In particular, ZoneBudget v3.01 or earlier
will NOT work with this option enabled.