test: example problem output update to reflect bug fixes
To keep track of changes in the code, a subset of the example
problems are kept in a "output-true" which should be nearly
identical to the results stored in the "output" directory.

To reflect the changes in the code from version
2.0 to 2.2, the output-true have been updated to
reflect the current state of the code.

The vast majority of differences offer after the
eighth significant digit and those that are
different beyond precision are due to bug fixes
that have occured since the release of version 2.0