test: automate example problem validation
Updated the examples/bash_example_run bash scripts
to automate running of example problems
and added:
   examples/bash_example_run/1_RunValidation.sh
and
   examples/bash_example_run/validate_example_results.f90

1_RunValidation.sh clears the example/*/output directories,
runs each of the examples,
compiles validate_example_results.f90 and then runs it.

validate_example_results.f90 is a standalone program and
is compiled using gfortran. It was tested with gfortran v10.2.0
however, it should work gcc-5 or newer.

The validate_example_results.f90 program is setup
to read the results stored in:
                            "output/*_CumHCHG.txt"
and compare it to the corresponding:
                       "output-true/*_CumHCHG.txt"

If the program finds that the files do not match
within its specified tolerance, then it returns
a BASH error code of 66