fix: fmp error when not using sfr
If FMP is used without SFR, then an index
error occured when FMP attempted to check
for runoff and delivery locations.
Previously, this was avoided with the WATERSOURCE
keyword specifying that SFR is not availible.

Now it automatically sets the surface water source
if SFR is not part of the simuation.

Also improved how WBS IDs are read in.
Now if any are <0 or >NWBS, the ID is set to 0