fix: fmp well linked to mnw2 copies mnw2 row and col
FMP supply wells that are simulated with the MNW2 required
the user to specify the row and column of the well in
both MNW2 and the FMP Supply Well block.

The FMP Supply Well block still must specify a row and column
but their value is not used. Instead, FMP copies over the
row and column specified for the MNW2 well it is linked too.