refactor: removed util.f warning about no cnstnt or iprn
Previously, if the read utilities:
U1DREL, U2DINT, U2DREL, or U2DDBL read a line
that did not include the CNSTNT or IPRN inputs,
then a warning was raised saying that they
were not found so it will set
CNSTNT = 1.0 and
IPRN = -1

The warning was removed and instead the following
is written to the list file as a normal input transcript:

"Reading for XYZ, autoset the scale factor CNSTNT to 1.0, and the print flag IPRN to -1"

where XYZ is the input being read.

This prevents spamming the WARNING and LIST files with
a notification that input is being set to a default value.