fix: expression_parser inline if conditional 2d array error
If the expression_parser solved an expression that is composed
of 1D or 2D arrays, then it would raise an index error
when it encountered an inline if statement.

This only effects the MULT package.
SLANG uses a newer equation parser code.

Inline IF is defined as follows:

IF[ COND, trueANS, falseANS ]

where
   COND is a conditional expression (must contain <, <=, >, >=)
   trueANS  is the result returned if COND is true
   falseANS is the result returned if COND is false
            it is optional to include falseANS,
            if not present and COND is false,
               then the result returned is NaN

For example, given
x = 10.
y = 5.

would have:

IF[ x + 3 > y, 2*x, 10*y ]

would return 2*x or 20

Another example is

IF[ A < -1e-6 & A > 1e-6, 0, B / A ]

would return 0 if A is near zero, otherwise returns B divided by A