feat: hob keyword not_observed_value
The HOB package checks at the start of the input
for the keyword NOT_OBSERVED_VALUE which is then
followed by a floating point number that defines
the value assigned for observations that are not
within the simulation timeframe.

Previously, these values were set to HOBDRY,
making it ambiguous if the observation point was
either from a cell that went dry or if it just was
not within the simulation timeframe.

The input location for the keyword is at the
start of the file. If there are multiple OWHM
HOB keywords in use, then only may be used per line.

For example:

--------------------------------------------
TIME_STEP_PRINT_ALL  obs.txt # obs.txt is has HOB output; is rewritten each time step (file is opened with Generic_Output, which supports differnet input styles, such as unit numbers or keyword directives and post-keywords).
NOT_OBSERVED_VALUE  -8888.   # Value if obs point is never observed
168190 5310 12 0    -9999.   # NH, MOBS, MAXM, IUHOBSV, HOBDRY
...
--------------------------------------------

In this example, the HOB file is written to obs.txt at the end of
each time step (note the file is overwritten after each time step).
(Note, obs.txt, is opened with Generic_Output)

Any observation points (ie their time of observation)
that occur before the simulation start or points that
are beyond the end of simulation or not yet simulated
are set to -8888.0

If an observation point is within the simulation timeframe,
but is within a cell that has gone dry (IBOUND = 0), then
it is set to -9999.0