fix: fmp surface_water block print sfr_return binary write error
If the FMP SURFACE_WATER block uses the option "PRINT SFR_RETURN"
and is set to write binary output, then it would attempt
to write the segment and reach for a WBS/Farm with
no assigned return flow locations.

Instead now, if the WBS/Farm does not have a return flow location,
then the segment and reach are set to zero.

Note this is already how the text version of the option already works.