fix: print_water_table and print_water_depth file header write error
In bas.f the subroutine GWF2BAS7OT checks if
PRINT_WATER_TABLE and/or PRINT_WATER_DEPTH options are in use and
if the corresponding arrays should be written to a file for
the stress period and time step that just finsihed.

While writting the header to the output, the code would use the
file unit number associated with PRINT_HEAD option, which would either
put an extra header in those files or raise a random access violation error.