fix: xmd_lib.f reorder declaring variables
Several routines in xmd_lib.f declare the
subroutine arguments after local variables.

Most compilers work around this, but it can
cause strange effects for arguments that represent
local array dimensions.

For example,
           subroutine array_routine(dim)
             integer, dimension(dim):: work
             integer:: dim
           end subroutine

Has the issue of dim being used before being declared.
Most compilers, will catch this and automatically use
dim appropiately, but it is  best to restructure
the routine as:
           subroutine array_routine(dim)
             integer:: dim
             integer, dimension(dim):: work
           end subroutine

or even better as:
           subroutine array_routine(dim)
             integer, intent(in):: dim
             integer, dimension(dim):: work
           end subroutine