fix: misc fmp files change 0d0 to 0.d0
Several of the FMP source files incorrectly
initialized a double precision variable using
the d0 syntax, but was missing the decimal point.

For example
   double precision:: X

   X = 0D0  ! is not correct
            !  -> 0D0 is an integer
   X = 0.D0 ! sets to double precision 0.0

This is a minor problem as most compilers can
convert the integer 0_int32 to
the double prececion 0.0_real64
on the fly.

This just ensures that the code comforms
to Fortran Standard.