doc: fmp_template remove nrd_infiltration_location keyword
In the FMP_Template, the input SURFACE_WATER included the option
NRD_INFILTRATION_LOCATION, which specified an infiltration locations
for excess non-routed delivery water. However this feature was not
fully implimented, so the keyword is removed to not mislead users
into thinking it is a current feature.

This option will be added back in once it is fully supported in FMP.