fix: swr reach ending assigned layer
SWR uses the start and ending elevations of each
reach to determine the layer that it resides in.
If the reach elevation was above the top of layer 1
or below the bottom of NLAY, then the assigned layer
was undefined. Instead now, it is set to layer 1 or NLAY,
if the elevation is above or below the model grid, respectively.