refactor: improved wel and mnw2 missing itmp warning
For the MNW2 and WEL packages, if the end of file is reached
or the input fails to read ITMP, then ITMP is assumed to be
zero for the rest of the simulation.

When this occurred a formal warning was made. This was changed to
being a minor message in the LISTing file with the worlds:
   don't panic
in the warning message.

It also indicates to the user that if they are using LINEFEED or
FMP-Link that the rates may be initialized outside of the package.