fix: swo index error when sorting nonexistent required flow variables
If Surface Water Operations (SWO) did not define in
Slang any required flow variables, then an index error occurs
when attempting to sort the require flow downstream flow order.

This fix is to make the code correct, however it did not effect
the simulation results when running the release version of the code.