feat: improved fmp semi-routed return by adding lost_runoff variable
Runoff that is generated from FMP either flows to SFR
or leaves the model as lost runoff. The SFR locations are
either defined explicitly as a semi-routed return (SRR) points
or automatically as fully-routed return (FRR). FRR just
searched for any SFR segments/reaches that reside in a
Water Balance Subregion (WBS) and then populates the SRR
pointer arrays.

If there were no SRR locations, then the original
setup would set

   SWF%SRRLOC(I)%N = 0  ! Size of RUNOFF for WBS I

and allocate:

   ALLOCATE( SWF%SRRLOC(I)%RUNOFF(1) ) ! Holds WBS I's runoff

The and that extra allocation contained any runoff that had no
where to go. If this situation occurred a warning is raised
to let the user know that runoff is leaving the model.

The code was improved by adding

   SWF%SRRLOC(I)%LOST_RUNOFF

which holds runoff that has nowhere to go.

In addition to this, the code was updated to allow defining SRR
locations with the segment = 0 to indicate that runoff is
supposed to leave the model. There is no limit for how many
SRR locations that remove water from the model, but it is
recommended to only have one per WBS.

For example, for WBS 4, the user may set 60% of the runoff
to go to segment 5, reach 1 and 40% of the runoff to leave
the model domain. The input would have the following format:

/-------------------------------------------------------------------------------\
| # NSFR_RETURN = 2                                                             |
| #                                                                             |
| SEMI_ROUTED_RETURN  STATIC  LIST  INTERNAL                                    |
| #                                                                             |
| # ISRR, WBS_ID, SEGMENT, REACH, [FRAC]                                        |
|     1     4     5         1      0.6   # 60% runoff applied to Seg 5, Reach 1 |
|     2     4     0         0      0.4   # 40% runoff removed from model domain |
|                                                                               |
\-------------------------------------------------------------------------------/