refactor: hydfmt single precision var are now double precision
HYDFMT reads the binary file produced by the Hydmod package
and generates readable text files.

The code relied on implicit naming and used single precision
variables for writing the data and calculating decimal years.
This lead to precision issues when dealing with models that
used seconds as a time unit or if the Hydmod binary was
written in double precision (precision truncation).

The code was modified to still read a single precision
binary, but all other variables either rely on the follow
statement at the start of each routine:

      USE, INTRINSIC:: ISO_FORTRAN_ENV, ONLY: REAL64
      IMPLICIT REAL(REAL64) (A-H,O-Z)

or the variable is directly declaired as:
      REAL(REAL64)