chore: merge batteries included library (bif) ac75415 commit
The Batteries Included Library (BiF) is located at:
code.usgs.gov/fortran/bif
Parts of the library are used by MODFLOW-OWHM and stored
in the src/bif_lib directory.

This commit merges current changes from BiF.
For more details about changes to BiF, please see
the changelog located at:
code.usgs.gov/fortran/bif/-/blob/main/CHANGELOG.md

This commit adds the optional argument "position"
to SUBROUTINE GENERIC_OPEN in the file
src/io/generic_open_interface.fpp

The argument specifies the POSITION in a Fortran open statement.

All Fortran OPEN(POSITION=) strings are accepted.
However it is recommended to only use:
    "ASIS",   or
    "REWIND", or
    "APPEND"

If the POSITION argument is not given,
then it is set to "REWIND".

This argument has no effect if the file is opened with
ACTION="WRITE" because the a new, empty file is created.