fix: fmp error when a crop with salinity irrigation flush is not in use
Fixed an fortran crash that occurred when a crop is marked as
having salinity flushing, but the crop is not included in the
stress period. For example Crop 3 has salinity flushing, but is not
part of the simulation until Stress Period 5, then a crash would occur
during stress period 1 because it was not present but has flush irrigation.
This has been fixed to automatically not include the flushing when the crop
is not simulated.