feat: fmp climate allows specifying direct_recharge multiple times
The Climate Block in FMP has the option DIRECT_RECHARGE that
specifies an NROW by NCOL array that is applied to deep percolation.
This is similar to the RCH package, except if there is a UZF link then
the deep percolation is applied to UZF.

Previously, FMP only allowed specifying DIRECT_RECHARGE once within
the CLIMATE Block. That has been relaxed to specify it as many times
as necessary to allow separating out the recharge arrays.