refactor: pandoc change argument from self-contained to embed-resources
Pandoc is used to convert markdown files to pdf.
A set of bash scripts
.util/pandoc/*.sh
are setup to convert the repository markdown files
to pdf-latex format. The bash scripts contained the
following deprecated argument:
   --self-contained
which suggested changing it to:
   --embed-resources

This commit changes the argument in the bash files to remove
the deprecated warning.