fix: fmp crop_data ixj input processing error
SUBROUTINE PARSE_CROP_ROW_COL in src/fmp/crop_data.f90
had two bugs.

The first one occurs when using multiple crops and
specifying the crop fraction of area with IXJ and
there are crop fractions set to zero. The loop
that searches the IXJ input had a hard limit
equal to the number of times the crop is specified,
but zero fractions are skipped causing a mismatch.
The code was modified to use a WHILE loop that terminates
when all the entire IXJ values have been searched.

The second is if when specifying single crop by row and column
with IXJ. The code determined the number of cells from the
Fraction IXJ rather than the cell ID. This resulted in
the code never loading IXJ specified crops that are defined
as one crop per model cell.