feat: updated cfp list directed read to owhm/bif read util
The input to CFP was updated to use the BiF get routines.
The input no longer requires comment lines, but any comment MUST
be preceded by a #.

The input was also adjusted to match the reported CFP official documentation.
Originally the input had an extra integer read that caused a mismatch with
official documentation. This was an accidental inclusion and now is removed.