test: nwt prob3 reduce rclose
The mf-nwt Prob3 example problems had
Pr3.nwt reduce the RCLOSE to improve
the repeatability of the problem among
different compiler versions.

Lose criteria resulted in owhm automatic tests
failing for different compilers.

This is just to ensure that the result is exactly
the same between different floating point models.