test: mf-2005 MNW2-Fig28 example reduce pcg rclose
The mf-2005 MNW2-Fig28 example problem reduced
the MNW2-Fig28.pcg RCLOSE to improve
the repeatability of the problem among
different compiler versions.

The larger criteria resulted in owhm automatic
tests failing for different compilers.

This is just to ensure that the result is exactly
the same between different floating point models.