feat: improve lpf/upw layer wettable description
The LPF/UPW packages output the layer wettable description as being
either "NEVER-DRY" or "WETTABLE", which is stored in the WETNAM variable.
To be more description the variable now outputs:
"NEVER-DRY", "STAY-DRY", or "WETTABLE".

"NEVER-DRY" - Indicates layer is confine
"STAY-DRY"  - Indicates layer is unconfine without wetting
"WETTABLE"  - Indicates layer is unconfine with    wetting