fix: nwt input allows continue option after specify
The NWT solver instruction defines that a set of
numbers are read after the SPECIFY keyword option.
However, the MODFLOW-NWT source code allows for
the CONTINUE keyword option to appear after SPECIFY
and before the numerical parameters. This is also
the default behavior for FloPy when making a NWT
solver file. To maintain compatibility with
MODFLOW-NWT and FloPy, this commit adds a check
for the CONTINUE keyword after SPECIFY.