feat: add command arguments -h and -v
The command arguments
-h and --help
-v and --version
have been added.

If they are invoked, then general information is printed
and then the standard header that includes the version is printed
or just the standard header with the version.

If a - or -- argument is present then the simulation
does not run beyond what the argument does.