fix: slang add check for valid req.delivery.farm variables
Slang property variables are set to values within OWHM
based on the current model state. The variables
REQ.DELIVERY.FARM.# and REQ.DELIVERY_VOL.FARM.#
where # is replaced by a valid farm/wbs number.

However if the user entered a bad number,
then either the variable was set to a random
value or resulted in a runtime index error.