fix: swo adapts when to update slang sfr.inflow variables
Slang property variables are set to values within OWHM
based on the current model state. The variables
SFR.seg#.INFLOW and SFR.seg#.OUTFLOW
are set only after the first solver iteration.
(seg# is the SFR segment number.)

This is to represent the flow in the SFR
network without the influence of SWO.

If a decision needs to be made based on SWO
release-induced flow, then slang must be run
at the solver iteration level. This fix
checks if an iteration level slang script
is evaluated, and if true, updates the
inflow and outflow variables every iteration.