fix: slang incorrectly handled variable name duplicate errors
Slang variable names are supposed to be unique. However,
if the RETURN and PROPERTY blocks contained a variable
with the same name, such as each having SFR.16.INFLOW,
then the slang is supposed to remove the PROPERTY block
duplicate. However, it would remove instead the last
variable in the PROPERTY block. For each duplicate,
would result in another variable being removed.