fix: remove pure statement from routines that are not truely pure
The PURE statement is preferenced before subroutines and functions
to indicate there are no side effects and the compiler can parallelize
calls to the routines.

Several routines were previously not flagged by the compiler
as being invalid for the PURE statement. They have been removed
for changed to IMPURE for routines that do not meet the defintion
of a pure routine. This fix allows for compliation with
Intel oneAPI v2024.2.0 released June 18, 2024.