refactor: examples/bash_example_run/1_RunValidation.sh error committed
The example driver bash script runs all the example problems and
then runs a fortran tool to check if they match a known hash value.
If one or more examples do not match, then the fortran tool returns
a non-zero exit status code. However, on some versions of bash
this would instead exit the entire script preventing the
closing clean up and comments from running. Instead a set
of if and conditional || are used to check if the examples
pass and then the script completes normally.