feat: fmp no longer stops when using srd/srr without sfr
The FMP previously would raise an error if the SFR
package was not declared the the name file and
the SURFACE_WATER block input included one of the
following options: semi-routed delivery points,
semi-routed  return points, or fully-routed return points.

Now the program raises a warning and disables those options,
since they have no meaning when SFR is not enabled.