fix: swo unable to read absolute_convergence_criteria option
The Surface Water Operations input contained a bug
with the arguments passed to ULOAD when parsing the
ABSOLUTE_CONVERGENCE_CRITERIA and RELATIVE_CONVERGENCE_CRITERIA
keywords. If they keywords were specified, then an error
saying "Unable to open EXTERNAL XYZ", where XYZ was a
random, small integer. This problem arose because the
ULOAD unit number must be set to zero to indicate a
new file or unit number must be specified, so the last value
stored in that number was passed in resulting in the error message.