refactor: changed sfr reach depth error to warning
If an SFR reach has a depth lower then the model's cells
bottom, then an error is raised an the program stopped. However,
this creates an issue when SFR adjusts the layer higher because
the layer above initially is active. This leaves the user with
no options other than setting the SFR layer to a higher elevation,
which may not be correct. Instead, this error was changed to a warning
with suggestions for fixing the issue.