feat[BREAKING]: fmp fnr in cbc stores et and dperc separately
Previously, the FMP writes the farm net recharge (FNR) to the
Cell-By-Cell (CBC) as a NROW by NCOL array that stores the
difference between deep percolation (DPERC) and evapotranspiration
from groundwater (ETgw), that is the CBC stores only `DPERC-ETgw`.

This feature modified the FMP CBC writer to write instead
the layer, row, column and flow rate (WEL-like-CBC record)
for DPERC and ETgw separately. This allows post-processing
programs, such as ZoneBudget, to accumulate budgets with
the IN term only containing DPERC and the OUT term to
only contain ETgw.

This change makes the IN and OUT match the LISTing
file Volumetric Budget IN and OUT terms for the FNR.

This is a breaking change in that the CBC will produce
a different IN and OUT for the same input, but the
net IN-OUT, is the same.

The FMP OUTPUT block added the option:
`COMPOSITE_FARM_NET_RECHARGE_CBC`
to write the original FNR CBC format, `DPERC-ETgw`.