fix: drt automatic_negative_itmp option error
The DRT package option AUTOMATIC_NEGATIVE_ITMP is supposed
to indicate that the first stress period is read, then
the remaining stress periods are assumed to have ITMP = -1
but instead it set all stress periods to -1, disabling the
DRT package. This fix allows reading the first stress period
and then reuses it for the remainder of the simulation.