fix: lmt-sfr link contained pack function bug
The LMT routine LMT8SFR2 attempts to pull all the non-empty
strings stored in FLOWTYPE by using the PACK function and
a MASK. the problem is that it is stored in
CHARACTER(16), DIMENSION(5):: PRNTSFRQTYP
so if the MASK is not all TRUE then a runtime error happens.

A simple fix would be to change:
    `PRNTSFRQTYP = PACK(FLOWTYPE, MASK)`
to:
    `PRNTSFRQTYP(:count(MASK)) = PACK(FLOWTYPE, MASK)`

Instead the entire MASK search code was replaced bu a simple loop
that removed the need for the PACK function call. This refactoring
removed about 100 lines of unnecessary code.