fix: lmt set nflowtype after it is written
Two fixes in the commit. The first is a block
of code is commented out that increased a counter
for the advanced packages. While the code itself is
fine, it also requires that MT3D be modified to
expdect input from the advanced packages.

The second fix is that the SFR flag NFLOWTYPE was
set after it was written. This code has been moved
to before its write to now write out the correct value.